refactor(resources): apply service/repository pattern to blob, collection, and song
Extracts blob, collection, and song from the old two-file (model.rs + rest.rs)
pattern into the same three-layer architecture introduced for setlist:

- model.rs: XxxRecord only (no DB impl, no _for_user methods)
- repository.rs: XxxRepository trait (pure data access, pre-resolved teams)
- surreal_repo.rs: SurrealXxxRepo implementing the trait against SurrealDB
- service.rs: XxxService<R, T, ...> orchestrating repo + TeamResolver + extras
- rest.rs: handlers now inject Data<XxxServiceHandle> instead of Data<Database>

Blob additionally introduces a BlobStorage trait + FsBlobStorage to decouple
filesystem I/O from DB access. Song introduces UserCollectionUpdater to
abstract the default-collection side-effect in create_song_for_user.
CollectionRepository.add_song_to_collection is reused by SongService for
the same side-effect, breaking the direct collection/song model coupling.

Integration tests migrate from model.rs into service.rs test modules;
test_helpers gains blob_service(), collection_service(), and song_service()
factory functions matching the existing setlist_service() pattern.

All 35 tests pass; cargo clippy is clean.

Made-with: Cursor
