+++ backend/src/resources/blob/mod.rs
@@ -1,7 +1,14 @@
 pub use shared::blob::{Blob, CreateBlob};
 
 mod model;
-#[allow(unused_imports)]
-pub use model::Model;
+mod repository;
+pub mod service;
+mod surreal_repo;
+pub mod storage;
+
+pub use repository::BlobRepository;
+pub use service::{BlobService, BlobServiceHandle};
+pub use surreal_repo::SurrealBlobRepo;
+pub use storage::FsBlobStorage;
 
 pub mod rest;
