+++ backend/src/resources/collection/repository.rs
@@ -1,0 +1,56 @@
+use async_trait::async_trait;
+use surrealdb::sql::Thing;
+
+use shared::api::ListQuery;
+use shared::collection::{Collection, CreateCollection};
+use shared::song::{Link as SongLink, LinkOwned as SongLinkOwned};
+
+use crate::error::AppError;
+
+/// Pure collection data access (no user ACL — callers pass pre-resolved team [`Thing`]s).
+#[async_trait]
+pub trait CollectionRepository: Send + Sync {
+    async fn get_collections(
+        &self,
+        read_teams: Vec<Thing>,
+        pagination: ListQuery,
+    ) -> Result<Vec<Collection>, AppError>;
+
+    async fn get_collection(
+        &self,
+        read_teams: Vec<Thing>,
+        id: &str,
+    ) -> Result<Collection, AppError>;
+
+    async fn get_collection_songs(
+        &self,
+        read_teams: Vec<Thing>,
+        id: &str,
+    ) -> Result<Vec<SongLinkOwned>, AppError>;
+
+    async fn create_collection(
+        &self,
+        owner: &str,
+        collection: CreateCollection,
+    ) -> Result<Collection, AppError>;
+
+    async fn update_collection(
+        &self,
+        write_teams: Vec<Thing>,
+        id: &str,
+        collection: CreateCollection,
+    ) -> Result<Collection, AppError>;
+
+    async fn delete_collection(
+        &self,
+        write_teams: Vec<Thing>,
+        id: &str,
+    ) -> Result<Collection, AppError>;
+
+    async fn add_song_to_collection(
+        &self,
+        write_teams: Vec<Thing>,
+        id: &str,
+        song_link: SongLink,
+    ) -> Result<(), AppError>;
+}
