+++ backend/src/resources/collection/rest.rs
@@ -3,7 +3,6 @@ use actix_web::{
     web::{self, Data, Json, Path, Query, ReqData},
 };
 
-use crate::database::Database;
 #[allow(unused_imports)]
 use crate::docs::ErrorResponse;
 use crate::error::AppError;
@@ -11,6 +10,7 @@ use crate::resources::User;
 #[allow(unused_imports)]
 use crate::resources::collection::Collection;
 use crate::resources::collection::CreateCollection;
+use crate::resources::collection::service::CollectionServiceHandle;
 #[allow(unused_imports)]
 use crate::resources::song::{Format, QueryParams, Song};
 use shared::api::ListQuery;
@@ -50,12 +50,12 @@ pub fn scope() -> Scope {
 )]
 #[get("")]
 async fn get_collections(
-    db: Data<Database>,
+    svc: Data<CollectionServiceHandle>,
     user: ReqData<User>,
     query: Query<ListQuery>,
 ) -> Result<HttpResponse, AppError> {
     Ok(HttpResponse::Ok().json(
-        db.list_collections_for_user(&user, query.into_inner())
+        svc.list_collections_for_user(&user, query.into_inner())
             .await?,
     ))
 }
@@ -81,11 +81,11 @@ async fn get_collections(
 )]
 #[get("/{id}")]
 async fn get_collection(
-    db: Data<Database>,
+    svc: Data<CollectionServiceHandle>,
     user: ReqData<User>,
     id: Path<String>,
 ) -> Result<HttpResponse, AppError> {
-    Ok(HttpResponse::Ok().json(db.get_collection_for_user(&user, &id).await?))
+    Ok(HttpResponse::Ok().json(svc.get_collection_for_user(&user, &id).await?))
 }
 
 #[utoipa::path(
@@ -109,11 +109,11 @@ async fn get_collection(
 )]
 #[get("/{id}/player")]
 async fn get_collection_player(
-    db: Data<Database>,
+    svc: Data<CollectionServiceHandle>,
     user: ReqData<User>,
     id: Path<String>,
 ) -> Result<HttpResponse, AppError> {
-    Ok(HttpResponse::Ok().json(db.collection_player_for_user(&user, &id).await?))
+    Ok(HttpResponse::Ok().json(svc.collection_player_for_user(&user, &id).await?))
 }
 
 #[utoipa::path(
@@ -143,12 +143,12 @@ async fn get_collection_player(
 )]
 #[get("/{id}/export")]
 async fn get_collection_export(
-    db: Data<Database>,
+    svc: Data<CollectionServiceHandle>,
     user: ReqData<User>,
     id: Path<String>,
     query: Query<QueryParams>,
 ) -> Result<HttpResponse, AppError> {
-    db.export_collection_for_user(&user, &id, query.into_inner().format)
+    svc.export_collection_for_user(&user, &id, query.into_inner().format)
         .await
 }
 
@@ -173,11 +173,11 @@ async fn get_collection_export(
 )]
 #[get("/{id}/songs")]
 async fn get_collection_songs(
-    db: Data<Database>,
+    svc: Data<CollectionServiceHandle>,
     user: ReqData<User>,
     id: Path<String>,
 ) -> Result<HttpResponse, AppError> {
-    Ok(HttpResponse::Ok().json(db.collection_songs_for_user(&user, &id).await?))
+    Ok(HttpResponse::Ok().json(svc.collection_songs_for_user(&user, &id).await?))
 }
 
 #[utoipa::path(
@@ -198,12 +198,12 @@ async fn get_collection_songs(
 )]
 #[post("")]
 async fn create_collection(
-    db: Data<Database>,
+    svc: Data<CollectionServiceHandle>,
     user: ReqData<User>,
     payload: Json<CreateCollection>,
 ) -> Result<HttpResponse, AppError> {
     Ok(HttpResponse::Created().json(
-        db.create_collection_for_user(&user, payload.into_inner())
+        svc.create_collection_for_user(&user, payload.into_inner())
             .await?,
     ))
 }
@@ -230,13 +230,13 @@ async fn create_collection(
 )]
 #[put("/{id}")]
 async fn update_collection(
-    db: Data<Database>,
+    svc: Data<CollectionServiceHandle>,
     user: ReqData<User>,
     id: Path<String>,
     payload: Json<CreateCollection>,
 ) -> Result<HttpResponse, AppError> {
     Ok(HttpResponse::Ok().json(
-        db.update_collection_for_user(&user, &id, payload.into_inner())
+        svc.update_collection_for_user(&user, &id, payload.into_inner())
             .await?,
     ))
 }
@@ -262,9 +262,9 @@ async fn update_collection(
 )]
 #[delete("/{id}")]
 async fn delete_collection(
-    db: Data<Database>,
+    svc: Data<CollectionServiceHandle>,
     user: ReqData<User>,
     id: Path<String>,
 ) -> Result<HttpResponse, AppError> {
-    Ok(HttpResponse::Ok().json(db.delete_collection_for_user(&user, &id).await?))
+    Ok(HttpResponse::Ok().json(svc.delete_collection_for_user(&user, &id).await?))
 }
