+++ backend/src/resources/song/mod.rs
@@ -2,11 +2,15 @@ pub use shared::song::{CreateSong, Song};
 
 mod liked;
 mod model;
-#[allow(unused_imports)]
-pub use model::Model;
-pub use model::SongRecord;
+mod repository;
+pub mod service;
+mod surreal_repo;
 
 pub use liked::LikedSongIds;
+pub use model::SongRecord;
+pub use repository::SongRepository;
+pub use service::{SongService, SongServiceHandle};
+pub use surreal_repo::SurrealSongRepo;
 
 pub mod rest;
